Introduction to DataFrames in Pandas
Welcome to Day 1 of our 45-day Data Science with AI Challenge!

Today, we’ll learn about DataFrames — one of the most important data structures in Python for data analysis.

💡 What is a DataFrame?
A DataFrame is like a table — with rows and columns, just like an Excel sheet.

It comes from the Pandas library, which is a powerful tool for working with data in Python. It has functions for analysing, cleaning, exploring, and manipulating data.